Rename things and reorganize to prepare for dual asyncio and trio support
