+++ docs/aio_run_in_process.rst
@@ -1,53 +1,53 @@
-asyncio\_run\_in\_process package
+aio\_run\_in\_process package
 =================================
 
 Submodules
 ----------
 
-asyncio\_run\_in\_process.abc module
+aio\_run\_in\_process.abc module
 ------------------------------------
 
-.. automodule:: asyncio_run_in_process.abc
+.. automodule:: aio_run_in_process.abc
     :members:
     :undoc-members:
     :show-inheritance:
 
-asyncio\_run\_in\_process.exceptions module
+aio\_run\_in\_process.exceptions module
 -------------------------------------------
 
-.. automodule:: asyncio_run_in_process.exceptions
+.. automodule:: aio_run_in_process.exceptions
     :members:
     :undoc-members:
     :show-inheritance:
 
-asyncio\_run\_in\_process.process module
+aio\_run\_in\_process.process module
 ----------------------------------------
 
-.. automodule:: asyncio_run_in_process.process
+.. automodule:: aio_run_in_process.process
     :members:
     :undoc-members:
     :show-inheritance:
 
-asyncio\_run\_in\_process.run\_in\_process module
+aio\_run\_in\_process.run\_in\_process module
 -------------------------------------------------
 
-.. automodule:: asyncio_run_in_process.run_in_process
+.. automodule:: aio_run_in_process.run_in_process
     :members:
     :undoc-members:
     :show-inheritance:
 
-asyncio\_run\_in\_process.state module
+aio\_run\_in\_process.state module
 --------------------------------------
 
-.. automodule:: asyncio_run_in_process.state
+.. automodule:: aio_run_in_process.state
     :members:
     :undoc-members:
     :show-inheritance:
 
-asyncio\_run\_in\_process.typing module
+aio\_run\_in\_process.typing module
 ---------------------------------------
 
-.. automodule:: asyncio_run_in_process.typing
+.. automodule:: aio_run_in_process.typing
     :members:
     :undoc-members:
     :show-inheritance:
@@ -56,7 +56,7 @@ asyncio\_run\_in\_process.typing module
 Module contents
 ---------------
 
-.. automodule:: asyncio_run_in_process
+.. automodule:: aio_run_in_process
     :members:
     :undoc-members:
     :show-inheritance:
