+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# asyncio-run-in-process documentation build configuration file, created by
+# aio-run-in-process documentation build configuration file, created by
 # sphinx-quickstart on Thu Oct 16 20:43:24 2014.
 #
 # This file is execfile()d with the current directory set to its
@@ -53,7 +53,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = 'asyncio-run-in-process'
+project = 'aio-run-in-process'
 copyright = '2019, The Ethereum Foundation'
 
 __version__ = setup_version
@@ -189,7 +189,7 @@ html_static_path = ['_static']
 #html_file_suffix = None
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'asyncio_run_in_processdoc'
+htmlhelp_basename = 'aio_run_in_processdoc'
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -209,7 +209,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-  ('index', 'asyncio_run_in_process.tex', 'asyncio-run-in-process Documentation',
+  ('index', 'aio_run_in_process.tex', 'aio-run-in-process Documentation',
    'The Ethereum Foundation', 'manual'),
 ]
 
@@ -239,7 +239,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', 'asyncio_run_in_process', 'asyncio-run-in-process Documentation',
+    ('index', 'aio_run_in_process', 'aio-run-in-process Documentation',
      ['The Ethereum Foundation'], 1)
 ]
 
@@ -253,8 +253,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', 'asyncio-run-in-process', 'asyncio-run-in-process Documentation',
-   'The Ethereum Foundation', 'asyncio-run-in-process', 'Simple asyncio friendly replacement for multiprocessing',
+  ('index', 'aio-run-in-process', 'aio-run-in-process Documentation',
+   'The Ethereum Foundation', 'aio-run-in-process', 'Simple async friendly replacement for multiprocessing for asyncio or trio',
    'Miscellaneous'),
 ]
