+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist=
-    py{36,37,py3}-core
+    py{36,37,py3}-{core,trio,asyncio}
     lint
     doctest
 
@@ -9,7 +9,7 @@ combine_as_imports=True
 force_sort_within_sections=True
 include_trailing_comma=True
 known_third_party=hypothesis,pytest
-known_first_party=asyncio_run_in_process
+known_first_party=aio_run_in_process
 line_length=21
 multi_line_output=3
 use_parentheses=True
@@ -23,6 +23,8 @@ ignore=
 usedevelop=True
 commands=
     core: pytest {posargs:tests/core}
+    asyncio: pytest {posargs:tests-asyncio}
+    trio: pytest {posargs:tests-trio}
     doctest: make -C {toxinidir}/docs doctest
 basepython =
     doctest: python
@@ -32,13 +34,15 @@ basepython =
 extras=
     test
     doctest: doc
+    asyncio: test-asyncio
+    trio: test-trio
 whitelist_externals=make
 
 [testenv:lint]
 basepython=python
 extras=lint
 commands=
-    mypy -p asyncio_run_in_process --config-file {toxinidir}/mypy.ini
-    flake8 {toxinidir}/asyncio_run_in_process {toxinidir}/tests
-    isort --recursive --check-only --diff {toxinidir}/asyncio_run_in_process {toxinidir}/tests
-    pydocstyle {toxinidir}/asyncio_run_in_process {toxinidir}/tests
+    mypy -p aio_run_in_process --config-file {toxinidir}/mypy.ini
+    flake8 {toxinidir}/aio_run_in_process {toxinidir}/tests
+    isort --recursive --check-only --diff {toxinidir}/aio_run_in_process {toxinidir}/tests
+    pydocstyle {toxinidir}/aio_run_in_process {toxinidir}/tests
